Limit entities for tame task to those deriving from EntityTameable.
This could introduce issue with tameable mobs from other mods which do
not derive from EntityTameable. To be honest, this is an issue with the
mod: don't reinvent the wheel.

Additionally, change this to IEntityOwnable in case they DO implement
that (which is relevant), although it cannot be assumed the events will
properly fire for those.