Use entity ResourceLocation for tame tasks.
For some reason the EntityEntry can return "Ozelot" and that isn't the
ResourceLocation. However, as we're iterating over the map of
EntityEntries and ResourceLocations, it's a simple matter of storing the
string of the location instead of the name of the entry.

This could potentially be applied in the same way to the killing task
but I'm not sure if that would have any impact on current quests.