Taming task (#412)
* Animal Taming task.

It's basically the kill task except it's for tames. Also, the mob list
doesn't filter out untameable mobs yet.

* Resolve issues with String->UUID changes in old code.

Still to resolve is the EventTrigger for Animal Taming.

* Properly implement the EventTrigger for Animal Taming.

* Limit entities for tame task to those deriving from EntityTameable.

This could introduce issue with tameable mobs from other mods which do
not derive from EntityTameable. To be honest, this is an issue with the
mod: don't reinvent the wheel.

Additionally, change this to IEntityOwnable in case they DO implement
that (which is relevant), although it cannot be assumed the events will
properly fire for those.

* Actually implement checks for IEntityOwnable.

I'm still not convinced that it'll be enough, but it's enough for now.

* Restore missing adapter link that was accidentally deleted during merge.

* Case in point: AbstractHorse does neither EntityTameable or IEntityOwnable.

Also, fix the translation key for taming as the name of it is still
"reqKills".

* Modify translations for Edit Mob.

This menu function can be used for both killing and taming tasks.

* Use entity ResourceLocation for tame tasks.

For some reason the EntityEntry can return "Ozelot" and that isn't the
ResourceLocation. However, as we're iterating over the map of
EntityEntries and ResourceLocations, it's a simple matter of storing the
string of the location instead of the name of the entry.

This could potentially be applied in the same way to the killing task
but I'm not sure if that would have any impact on current quests.

* On second thoughts, just check EntityLivingBase.

If the animal isn't tameable or it doesn't generate the correct Tame
Event...