Allow new lines to be placed in entries using \n.
Line breaks are converted into "\n"s when editing an entry that contains
them, and then they are converted back when saving. Entering "\n"s into
the entry will thus present line breaks.

This is one of the first things I wanted when I started playing around
with HQM, and it's something I've seen a lot of other people ask for.