Resolve #424, #427: manual detection & Advancement tasks.
Tested in single-player and in a dedicated client/server to make sure
everything is working properly.

Consumption tasks which previously used the allowManua() boolean method
now override that method, which is defined in QuestTask with a default
of false.

Detection tasks now check for allowDetect (also defined in QuestTask as
per above, with a default of false), instead of relying on instance of
Item Detection tasks.

Advancement task: use the "Location" mode button to change an
Advancement task, and select from any of the registered Advancements.

The task triggers either manually, or whenever the player achieves that
specific advancement.