Resolve #434: better visual feedback for "Select Quest" in Edit Mode.
The "selected" quest will now specifically show up in a different colour
from other quests, and different to the "currently being modified quest"
(ala changing parents, etc).

In addition, clicking on the same quest again will reset it.

Finally, I've renamed selectedQuestId to speciallySelectQuestId to better
clarify what it's actually used for and make it easier to look through
the code: it's easy to confuse it with other variables which are also
named "selectedQuestId".

I've updated the relevant tile entities to use this value instead.