Fix some missing null checks, properly finalise and initialise Quests.
I ran into the strangest concurrent modification error while testing the
new QuestCompletedTask that may have been a combination of a null
pointer error from another part of my code. However, I've properly set
it as final and initialised it as such.

Where it was previously initialised, I've simply cleared the contents.

I've also added a localisation for the questTrack bindTo as when I was
testing the changes made in the previous commit
(specificallySelectedQuestId) I noticed that I just got the translation
key rather than any actual message.

Finally, I added a null check to the getQuest static function to return
null instead of trying to consult the ConcurrentHashMap. I'm not sure if
it was related to a bug or if it's superfluous. What's the standard
behaviour for consulting a map with a null key?