Resolve #433: "(Other) quest completed" task.
This task fires whenever another quest has been completed. To use, use
the "Select Quest" option to select the relevant quest, then go to the
task and use either the: Edit Location, Change Name or Change Item
buttons to click on the task.

The task will then automatically fill in with the icon usesd for the
quest in question, and will fill in its name from the name of the quest.
It is not possible to overwrite either of these.

In the weird eventuality that somehow the quest in question is completed
and the task does not detect this, there is a manual detect button.