Resolve #431: Anvil "crafting" is detected, as is Furnace smelting.
Thus, a crafting recipe for cooked fish will now be marked as complete
if you take cooked fish out of the right-hand slot of a furance.

For Anvil recipes, enchanting a specific item with a specific enchant
and having precise detection on will allow you to detect that enchant.
For example, you can have a quest that requires you to "Craft" a simple
bow with an infinity enchant. Doing so meets the requirement.

However, enchanting a bow already enchanted with power will not meet the
requirement. I'll work on this presently.

Additionally, any recipe added by such things as RockyCore (extension of
CraftTweaker) seem to work as intended.