Resolve #376: required completed parents are now properly saved.
Previously they were only being saved if the "use modified parent count"
variable was true, which was only being set by the interface option to
change them, which meant that in the rewrite to file before load, the
value was never actually written and eventually got discarded.

(The fact that we have two rewrites before the game actual loads in is
something else that disturbs me.)

Likewise, the interface didn't function properly because whenever you
opened up the quest with the "required parents" tool it was looking for
that same variable.

I removed the boolean and just made it a function that checks to see if
the required parent count is less than the total size of parents.

This now works through multiple saves, editing via the usual tool (which
sets the arrow to the correct screen and sets the text box to the
correct value). It also works in a testing scenario with 3 parent quests
and a requirement of 2, and only completing 2 parents.

Additionally, it works in a situation where two of the parent quests are
linked together as optional: you can complete one of those and the other
quest, and then the final quest is unlocked.