Resolve #436: no issue with additional new parents, but issues with resetting to default.
This appears to be a disparity onArrowClick specifically setting the
boolean value but the boolean value not being considered as that boolean
value is no longer a member of the quest class.

Instead, when the boolean value is true, use the required parent count
from the text box, otherwise use the default -1.