Improve QuestTaskCompleted, improve interface and instructions for it.
Additionally, you can do any non-delete action (even just a simple
click) to set the quest type. However, you can only set it for those
tasks which have the relevant red text next to them.

It's possible to enable continuous setting (in order to change it), but
I think that makes it easy to mis-click and make mistakes.