Change /hqm edit, automatically disable edit-mode in multi-player.
Resolves #464: due to the effects of the reach-around between server and
client while in Integrated mode (ew), it was possible previously to be
on a dedicated server with all of the edit capabilities -- which did
nothing.

Also changes some commands:
 - /hqm edit -> /hqm op
 - /hqm mode -> /hqm edit

It makes no sense for the edit mode toggle to be on "mode" while having
another command called "edit", which in reality is just a command to get
the op book.

There's a reminder message for operators who use /hqm edit in a server
context of the new command they should use. I also moved almost all of
the text into translations.