Partially resolve #459, resolve #469, resolve #461.
Actually fire book events; provide a message that's more accurate than
"you have no permission" when the case may be that it's simply still
loading the book data.

I'm not sure what's causing the "That player doesn't exist" line in the
screenshot I posted; it seems unrelated.

Finally, prevent enabling edit mode while server sync is active with a
message as to why.