Resolve #478: clear invalid quest triggers on log-in.
It's entirely possible and extremely likely that quests from another
single-play game still have their tasks registered with the
EventTrigger's list of tasks after switching games.

PlayerLoggedIn is too late to simply clear the slate, but it is still
possible to check if a quest task is valid (i.e., is actually contained
within its parent).